+++ lib/gems/errors/network_error.rb
@@ -1,0 +1,6 @@
+require_relative "error"
+
+module Gems
+  # Error raised when a network error occurs
+  class NetworkError < Error; end
+end